+++ internal/cmds/version.go
@@ -1,4 +1,4 @@
-package main
+package cmds
 
 import (
 	"fmt"
@@ -8,7 +8,7 @@ import (
 
 var version = "unknown"
 
-func newVersionCmd() *cobra.Command {
+func NewVersionCmd() *cobra.Command {
 	return &cobra.Command{
 		Use:   "version",
 		Short: "Print the version number of Redfish Exporter",
