--- tools.go
@@ -1,8 +1,0 @@
-//go:build tools
-// +build tools
-
-package main
-
-import (
-	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
-)
